cancel future if a timeout occurs when polling
On java8 CompositeMetricPollerTest.testTimeout was
failing because the traversal order changed and the
task that timedout was done first. This used up the
pool and the subsequent futures all timed out because
they were waiting for a free thread to execute.
Switching the keys would cause it to pass on java8
and fail on java7.

This change cancels the future on timeout and updates
the test case to have several more timeout cases so
it will fail on both java7 and java8 if there is a
problem.