compute all stats based on capture values
Before there was inconsistency where some of the stats
like total and mean were being computed using all samples
passed to the `record` method, but others were only
using the samples in the rolling array. The variance
computation even mixed and used `count` in some places
and `curSize` in others.

This change makes all the stats consistently use the
set of values from the rolling array and ignore others.
This does mean that the total and mean will now only
reflect a subset of the values passed to `record`.