make StatsBuffer more resilient to overflow
Keeps track of the actual number of items that have
been recorded up to the buffer size. Before it was
using `min(count, size)`. If `count` overflowed, then
it could be negative or between `0` and `size` and
the wrong value would get used.

Note, this also changes the overall count to max out
at the buffer size. I also considered making `count`
a long, but that would require changing the public
signature of `getCount` which would break backwards
compatibility.