fix build error for RxHttp
Servo was using a long deprecated constructor in
RxHttp that was removed.

```
/home/travis/build/Netflix/servo/servo-atlas/src/main/java/com/netflix/servo/publish/atlas/AtlasMetricObserver.java:97: error: constructor RxHttp in class RxHttp cannot be applied to given types;
    this(config, commonTags, pollerIdx, new HttpHelper(new RxHttp(new BasicServerRegistry())));
                                                       ^
  required: Config,ServerRegistry
  found: BasicServerRegistry
  reason: actual and formal argument lists differ in length
1 error
:servo-atlas:compileJava FAILED
```

This change passes in an empty config. Note that
previously you could get the static config instance
from archaius1. With archaius2 there is no static
config, so if you are not injecting the configuration
you do not have any access to the configuration used by
the rest of the application.