add support to delegate to spectator
This is a first stab at adding a hook that will allow
updates that go to standard Servo types to be forwarded
to a Spectator registry. In most cases the updates will
be sent as they happen, e.g., the basic counter will
increment the spectator counter internally. This avoids
delays or other data quality issues.

Because so much of Servo is based on static, the main
integration point is a static reference in
SpectatorContext. The default registry is a noop
implementation so it must be set as early as possible
in the application setup.

To minimize the compatibility issues the existing
Servo implementations were changed as little as possible.
Concerns:

- Custom monitor implementations that get registered
  will currently get ignored. Will look into that later.
- User defined functions registered as gauges will now
  get called more frequently. This might cause load issues
  if they perform expensive operations inline.