fix registration tags in spectator integration (#448)
Before we would always set the spectator id based on the
initial config. For uses cases where a monitor field is
registered, the tags get update with the class and possibly
and id. Servo does this with a wrapper that changes the
id. For Spectator we need to be able to update the id and
fetch a new instance of the underlying meter that will get
updated.

Now the SpectatorMonitor interface allows the additional
tags to be set in an initialization call that happens as
part of the registration. The meter types are lazy so it
will not actually create them until the first activity
occurs which should be post-registration if they are using
Servo correctly.