do not warn if global registry used for context (#463)
If the Spectator global registry is used for the SpectatorContext,
then do not log a warning. In some legacy apps the context must
be set early to the global registry to avoid missing some metrics.
It can then be overwritten by the primary registry after it has
been created later in the startup process.