switch to no op monitor registry by default
Internally Servo is deprecated and delegates to Spectator.
This change would reduce the overhead for things that are
still using the Servo APIs. The main benefit is that the
data would no longer be exported to JMX.

Note, this could be quite disruptive to anyone else using
Servo. Since it is all setup via sytem properties, there
isn't a convenient way to only change the default we use
internally.

After making this change, I noticed that loading the JMX
monitor registry explicitly using the property no longer
worked. That has been fixed so the old behavior can be
acheived using a system property.